Veranderingen profit berekening
Profit wordt nu opgesplitst in drie delen, adhoc, parking en
reservation. Prijzen zijn apart in te stellen > verwerk in settings..

+++ Parkeersimulator/model/SimulatorView.java
@@ -249,5 +249,5 @@ public class SimulatorView {
         return simulator.getMissedCustomers();
     }
     public String getTime() {return simulator.getTime();}
-    public int getProfit() {return simulator.getProfit();}
+    public double getProfit() {return simulator.getProfit();}
 }
